[fix] enables to dump output files in original directory when restart

See merge request ricos/machine_learning/siml!241

+++ siml/trainer.py
@@ -34,13 +34,18 @@ class Trainer(siml_manager.SimlManager):
         """
 
         print(f"Output directory: {self.setting.trainer.output_directory}")
-        self.setting.trainer.output_directory.mkdir(parents=True)
+        overwrite_restart_mode = self._overwrite_restart_mode()
+        if not overwrite_restart_mode:
+            # When restart training, output directory already exists
+            self.setting.trainer.output_directory.mkdir(parents=True)
 
         self.prepare_training()
 
+        yaml_file_name = f"settings_restart_{util.date_string()}.yml" \
+            if overwrite_restart_mode else "settings.yml"
         setting.write_yaml(
             self.setting,
-            self.setting.trainer.output_directory / 'settings.yml',
+            self.setting.trainer.output_directory / yaml_file_name,
             key=self.setting.trainer.model_key)
 
         print(
@@ -54,16 +59,28 @@ class Trainer(siml_manager.SimlManager):
                 'validation/' + self._display_mergin(k)
                 for k in self.model.get_loss_keys()])
             + self._display_mergin('elapsed_time'))
-        with open(self.log_file, 'w') as f:
-            f.write(
-                'epoch, train_loss, '
-                + ''.join([
-                    f"train/{k}, " for k in self.model.get_loss_keys()])
-                + 'validation_loss, '
-                + ''.join([
-                    f"validation/{k}, " for k in self.model.get_loss_keys()])
-                + 'elapsed_time\n')
+        if not overwrite_restart_mode:
+            with open(self.log_file, 'w') as f:
+                f.write(
+                    'epoch, train_loss, '
+                    + ''.join([
+                        f"train/{k}, " for k in self.model.get_loss_keys()])
+                    + 'validation_loss, '
+                    + ''.join([
+                        f"validation/{k}, " for k in self.model.get_loss_keys()
+                    ])
+                    + 'elapsed_time\n')
         self.start_time = time.time()
+        if not overwrite_restart_mode:
+            self.offset_start_time = 0
+        else:
+            df = pd.read_csv(
+                self.log_file,
+                header=0,
+                index_col=None,
+                skipinitialspace=True
+            )
+            self.offset_start_time = df.tail(1).loc[:, "elapsed_time"].item()
 
         self.pbar = self.create_pbar(
             len(self.train_loader) * self.setting.trainer.log_trigger_epoch)
@@ -94,6 +111,16 @@ class Trainer(siml_manager.SimlManager):
             return train_state, validation_state, test_state
         else:
             return train_state, validation_state
+
+    def _overwrite_restart_mode(self) -> bool:
+        if self.setting.trainer.restart_directory is None:
+            return False
+
+        if self.setting.trainer.output_directory == \
+                self.setting.trainer.restart_directory:
+            return True
+
+        return False
 
     def _display_mergin(self, input_string, reference_string=None):
         if not reference_string:
@@ -324,7 +351,8 @@ class Trainer(siml_manager.SimlManager):
                 validation_other_losses = {}
             self.evaluation_pbar.close()
 
-            elapsed_time = time.time() - self.start_time
+            elapsed_time = (time.time() - self.start_time) \
+                + self.offset_start_time
 
             # Print log
             tqdm.write(
@@ -810,7 +838,6 @@ class Trainer(siml_manager.SimlManager):
             'max_epochs': self.setting.trainer.n_epoch,
             'epoch_length': len(self.train_loader),
         })
-        self.trainer.state.epoch = checkpoint['epoch']
         # self.loss = checkpoint['loss']
         print(f"{snapshot} loaded for restart.")
         return
